Refactor most read assertions into a single location to reduce duplication
